+++ lib/packages/wikibase/data-model/RELEASE-NOTES.md
@@ -6,6 +6,7 @@
 
 * The `Property` constructor now accepts an optional `StatementList` parameter
 * Added `Property::getStatements` and `Property::setStatements`
+* Added `StatementList::getAllSnaks` to use instead of `Entity::getAllSnaks`
 
 ## Version 1.0 (2014-09-02)