+++ lib/packages/wikibase/data-model/RELEASE-NOTES.md
@@ -8,6 +8,7 @@
 * Added `Property::getStatements` and `Property::setStatements`
 * Added `StatementList::getAllSnaks` to use instead of `Entity::getAllSnaks`
 * The `Statement` constructor now also accepts a `Claim` parameter
+* Added `Statement::setClaim`
 
 ## Version 1.0 (2014-09-02)