+++ lib/packages/wikibase/data-model/RELEASE-NOTES.md
@@ -7,6 +7,8 @@
 * The `Property` constructor now accepts an optional `StatementList` parameter
 * Added `Property::getStatements` and `Property::setStatements`
 * Added `StatementList::getAllSnaks` to use instead of `Entity::getAllSnaks`
+* The `Statement` constructor now also accepts a `Claim` parameter
+* Added `Statement::setClaim`
 
 ## Version 1.0 (2014-09-02)