+++ lib/packages/wikibase/data-model/RELEASE-NOTES.md
@@ -11,7 +11,7 @@
 * Added `Statement::setClaim`
 * Added `PropertyIdProvider` interface
 * Added `ByPropertyIdGrouper`
-* Added `BestStatementFinder`
+* Added `BestStatementsFinder`
 
 ## Version 1.0 (2014-09-02)