+++ lib/packages/wikibase/data-model/RELEASE-NOTES.md
@@ -12,6 +12,7 @@
 * Added `PropertyIdProvider` interface
 * Added `ByPropertyIdGrouper`
 * Added `BestStatementsFinder`
+* The `Reference` constructor now accepts a `Snak` array
 
 ## Version 1.0 (2014-09-02)